requirements.txt
tqdm==4.40.0
absl_py==0.8.1
bert_score==0.3.0
six==1.13.0
pyrouge==0.1.3
tensorflow==1.14.0
numpy==1.17.4
nltk==3.4.5